+++ tensorrt_llm/_torch/auto_deploy/custom_ops/__init__.py
@@ -9,6 +9,7 @@ from .linear import *
 from .mla import *
 from .quant import *
 from .torch_attention import *
+from .torch_moe import *
 from .torch_rope import *
 from .triton_attention import *
 from .triton_rope import *
